Convert the Bibliotheca integration into a package
The Bibliotheca integration was five flat bibliotheca_* modules sitting
directly in integration/license/, while every other integration
(overdrive/, boundless/, opds/) is a package. Restructure it to match.

The five modules become bibliotheca/{api,importer,purchase_record_importer,
circulation_updater,scripts}.py, and the 1295-line api.py is split along
the same seams the siblings use: settings.py (as in overdrive and
boundless), parser.py (as in boundless), and coverage.py (as in
overdrive).

Splitting the parsers out exposed a cycle: BibliothecaAPI constructs the
parsers, but ErrorParser, EventParser and PatronCirculationParser reached
back for BibliothecaAPI.SERVICE_NAME and .ARGUMENT_TIME_FORMAT. Add a
constants.py holding those values and point the parsers at it, mirroring
how boundless/parser.py and overdrive/coverage.py already avoid the same
cycle. BibliothecaAPI keeps both class attributes, so nothing that reads
them off the API changes.

Also drop EventParser.EVENT_SOURCE and EventParser.SET_DELIVERY_MECHANISM_AT.
Both were unread: EVENT_SOURCE has no readers anywhere, and the two sites
that read SET_DELIVERY_MECHANISM_AT do so off a BaseCirculationAPI, which
EventParser is not. Removing them lets parser.py stop importing the
circulation API layer entirely. BibliothecaAPI.SET_DELIVERY_MECHANISM_AT
is a different attribute on a different class and is untouched.

Mirror the layout in tests/, moving the fixture into a conftest.py as
boundless does. TestBibliographicCoverageProvider subclassed
TestBibliothecaAPI without using anything from it, which made pytest
re-run all 22 API tests a second time; it no longer does. All 38 declared
tests are preserved, with test_internal_formats moving to
TestItemListParser, the class it actually exercises.
