+++ bin/bibliotheca_circulation_update
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 """Manually kick off the Bibliotheca circulation update for one or all collections."""
-from palace.manager.integration.license.bibliotheca_scripts import (
+from palace.manager.integration.license.bibliotheca.scripts import (
     CirculationUpdateCollection,
 )
