+++ bin/bibliotheca_purchase_record_import
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 """Manually kick off the Bibliotheca purchase record import for one or all collections."""
 
-from palace.manager.integration.license.bibliotheca_scripts import (
+from palace.manager.integration.license.bibliotheca.scripts import (
     ImportPurchaseRecordCollection,
 )
