+++ bin/informational/bibliotheca-raw-bibliographic
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 from xml.dom import minidom
 
-from palace.manager.integration.license.bibliotheca import BibliothecaAPI
+from palace.manager.integration.license.bibliotheca.api import BibliothecaAPI
 from palace.manager.scripts.input import IdentifierInputScript
 from palace.manager.sqlalchemy.model.collection import Collection
