+++ bin/repair/bibliotheca_bibliographic_coverage
@@ -2,7 +2,7 @@
 """Make sure all Bibliotheca books have bibliographic coverage."""
 
 
-from palace.manager.integration.license.bibliotheca import (
+from palace.manager.integration.license.bibliotheca.coverage import (
     BibliothecaBibliographicCoverageProvider,
 )
 from palace.manager.scripts.coverage_provider import RunCollectionCoverageProviderScript
