+++ src/palace/manager/integration/license/bibliotheca/scripts.py
@@ -9,7 +9,7 @@ from sqlalchemy.orm import Session
 from palace.util.exceptions import PalaceValueError
 
 from palace.manager.celery.tasks import bibliotheca
-from palace.manager.integration.license.bibliotheca_purchase_record_importer import (
+from palace.manager.integration.license.bibliotheca.purchase_record_importer import (
     DEFAULT_PURCHASE_RECORD_START_TIME,
 )
 from palace.manager.scripts.base import Script
