+++ src/palace/manager/scripts/availability.py
@@ -43,13 +43,13 @@ class AvailabilityRefreshScript(IdentifierInputScript):
         collection = pool.collection
 
         if identifier.type == Identifier.BIBLIOTHECA_ID:
-            # Local import to avoid a circular import. bibliotheca_circulation_updater
+            # Local import to avoid a circular import. bibliotheca.circulation_updater
             # imports `palace.manager.celery.tasks`, whose package __init__ eagerly
             # autoloads every task module (including celery.tasks.bibliotheca), and
             # celery.tasks.bibliotheca imports BibliothecaCirculationUpdater back. A
             # top-level import here would trigger that cycle whenever availability.py is
             # imported before the Celery tasks have finished loading.
-            from palace.manager.integration.license.bibliotheca_circulation_updater import (
+            from palace.manager.integration.license.bibliotheca.circulation_updater import (
                 BibliothecaCirculationUpdater,
             )
