+++ tests/manager/scripts/test_configuration.py
@@ -4,7 +4,7 @@ from io import StringIO
 
 import pytest
 
-from palace.manager.integration.license.bibliotheca import BibliothecaAPI
+from palace.manager.integration.license.bibliotheca.api import BibliothecaAPI
 from palace.manager.integration.license.overdrive.api import OverdriveAPI
 from palace.manager.scripts.configuration import (
     ConfigureCollectionScript,
