+++ tests/manager/scripts/test_monitor.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pytest
 
 from palace.manager.core.monitor import CollectionMonitor, Monitor
-from palace.manager.integration.license.bibliotheca import BibliothecaAPI
+from palace.manager.integration.license.bibliotheca.api import BibliothecaAPI
 from palace.manager.integration.license.opds.opds1.api import OPDSAPI
 from palace.manager.scripts.monitor import (
     RunCollectionMonitorScript,
